# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real-Time Collaborative Code Editor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ExecuteCodeRequestLanguage =
  (typeof ExecuteCodeRequestLanguage)[keyof typeof ExecuteCodeRequestLanguage];

export const ExecuteCodeRequestLanguage = {
  python: "python",
  java: "java",
  cpp: "cpp",
  c: "c",
} as const;

export interface ExecuteCodeRequest {
  language: ExecuteCodeRequestLanguage;
  code: string;
}

export interface ExecuteCodeResponse {
  output: string;
  error?: string;
  exitCode: number;
}

export interface Version {
  id: string;
  username: string;
  code: string;
  language: string;
  timestamp: string;
  label?: string;
}

export interface SaveVersionRequest {
  username: string;
  code: string;
  language: string;
  label?: string;
}

export interface ErrorResponse {
  error: string;
}