# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real-Time Collaborative Code Editor API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Execute code in the specified language
 * @summary Execute code
 */
export const ExecuteCodeBody = zod.object({
  language: zod.enum(["python", "java", "cpp", "c", "javascript", "typescript"]),
  code: zod.string(),
});

export const ExecuteCodeResponse = zod.object({
  output: zod.string(),
  error: zod.string().optional(),
  exitCode: zod.number(),
});

/**
 * Returns list of saved code versions
 * @summary Get all versions
 */
export const GetVersionsResponseItem = zod.object({
  id: zod.string(),
  username: zod.string(),
  code: zod.string(),
  language: zod.string(),
  timestamp: zod.date(),
  label: zod.string().optional(),
});
export const GetVersionsResponse = zod.array(GetVersionsResponseItem);

/**
 * Save a snapshot of the current code
 * @summary Save a version
 */
export const SaveVersionBody = zod.object({
  username: zod.string(),
  code: zod.string(),
  language: zod.string(),
  label: zod.string().optional(),
});

/**
 * Returns a specific version by ID
 * @summary Get a specific version
 */
export const GetVersionParams = zod.object({
  id: zod.coerce.string(),
});

export const GetVersionResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  code: zod.string(),
  language: zod.string(),
  timestamp: zod.date(),
  label: zod.string().optional(),
});