# lib/api-zod/src/generated/types/executeCodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real-Time Collaborative Code Editor API
 * OpenAPI spec version: 0.1.0
 */
import type { ExecuteCodeRequestLanguage } from "./executeCodeRequestLanguage";

export interface ExecuteCodeRequest {
  language: ExecuteCodeRequestLanguage;
  code: string;
}