# lib/api-zod/src/generated/types/executeCodeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real-Time Collaborative Code Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteCodeResponse {
  output: string;
  error?: string;
  exitCode: number;
}